+++ src/main/java/shopping/application/CartItemService.java
@@ -7,9 +7,9 @@ import shopping.domain.cart.Product;
 import shopping.dto.request.CartItemCreateRequest;
 import shopping.dto.request.CartItemUpdateRequest;
 import shopping.dto.response.CartItemResponse;
-import shopping.exception.CartItemNotFoundException;
-import shopping.exception.ProductNotFoundException;
-import shopping.exception.UserNotMatchException;
+import shopping.exception.cart.CartItemNotFoundException;
+import shopping.exception.cart.ProductNotFoundException;
+import shopping.exception.auth.UserNotMatchException;
 import shopping.repository.CartItemRepository;
 import shopping.repository.ProductRepository;
