+++ src/main/java/shopping/auth/AuthArgumentResolver.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import shopping.exception.UserNotFoundException;
+import shopping.exception.auth.UserNotFoundException;
 import shopping.repository.UserRepository;
 
 import javax.servlet.http.HttpServletRequest;
