+++ src/main/java/shopping/auth/PBKDF2PasswordEncoder.java
@@ -1,6 +1,6 @@
 package shopping.auth;
 
-import shopping.exception.PasswordNotHashException;
+import shopping.exception.auth.PasswordNotHashException;
 
 import javax.crypto.SecretKeyFactory;
 import javax.crypto.spec.PBEKeySpec;
