--- src/main/java/shopping/exception/NoOrderItemException.java
@@ -1,9 +1,0 @@
-package shopping.exception;
-
-public class NoOrderItemException extends  ShoppingBaseException{
-    private static final int STATUS_CODE = 500;
-
-    public NoOrderItemException() {
-        super("", STATUS_CODE);
-    }
-}
