+++ src/main/java/shopping/exception/auth/PasswordNotHashException.java
@@ -1,6 +1,8 @@
-package shopping.exception;
+package shopping.exception.auth;
 
-public class PasswordNotHashException extends ShoppingBaseException{
+import shopping.exception.general.ShoppingBaseException;
+
+public class PasswordNotHashException extends ShoppingBaseException {
 
     private static final int STATUS_CODE = 500;
     private static final String MESSAGE = "패스워드 해시 과정에 에러가 발생하였습니다.";
