+++ src/main/java/shopping/exception/cart/NoOrderItemException.java
@@ -1,0 +1,11 @@
+package shopping.exception.cart;
+
+import shopping.exception.general.ShoppingBaseException;
+
+public class NoOrderItemException extends ShoppingBaseException {
+    private static final int STATUS_CODE = 500;
+
+    public NoOrderItemException() {
+        super("", STATUS_CODE);
+    }
+}
