+++ src/main/java/shopping/exception/cart/NotSameOrderException.java
@@ -1,4 +1,6 @@
-package shopping.exception;
+package shopping.exception.cart;
+
+import shopping.exception.general.ShoppingBaseException;
 
 public class NotSameOrderException extends ShoppingBaseException {
     private static final int STATUS_CODE = 500;
