+++ src/main/java/shopping/exception/cart/ProductNotFoundException.java
@@ -1,4 +1,6 @@
-package shopping.exception;
+package shopping.exception.cart;
+
+import shopping.exception.general.ShoppingBaseException;
 
 public class ProductNotFoundException extends ShoppingBaseException {
