+++ src/test/java/shopping/application/CartItemServiceTest.java
@@ -11,9 +11,9 @@ import shopping.domain.user.User;
 import shopping.dto.request.CartItemCreateRequest;
 import shopping.dto.request.CartItemUpdateRequest;
 import shopping.dto.response.CartItemResponse;
-import shopping.exception.CartItemNotFoundException;
-import shopping.exception.ProductNotFoundException;
-import shopping.exception.UserNotMatchException;
+import shopping.exception.cart.CartItemNotFoundException;
+import shopping.exception.cart.ProductNotFoundException;
+import shopping.exception.auth.UserNotMatchException;
 import shopping.repository.CartItemRepository;
 import shopping.repository.ProductRepository;
 import shopping.repository.UserRepository;
